Working skeleton with pingpong-bot
