--- .air.toml
@@ -1,32 +1,0 @@
-root = "."
-tmp_dir = "tmp"
-
-[build]
-  bin = "./tmp/main"
-  cmd = "go build -o ./tmp/main cmd/api/main.go"
-  delay = 1000
-  exclude_dir = ["assets", "tmp", "vendor"]
-  exclude_file = []
-  exclude_regex = []
-  exclude_unchanged = false
-  follow_symlink = false
-  full_bin = ""
-  include_dir = []
-  include_ext = ["go", "tpl", "tmpl", "html", ".env"]
-  kill_delay = "0s"
-  log = "build-errors.log"
-  send_interrupt = false
-  stop_on_error = true
-
-[color]
-  app = ""
-  build = "yellow"
-  main = "magenta"
-  runner = "green"
-  watcher = "cyan"
-
-[log]
-  time = false
-
-[misc]
-  clean_on_exit = false
