+++ .gitignore
@@ -25,3 +25,4 @@
 .env
 tmp
 vendor
+.vscode
