+++ bot/domain/discord_commands.go
@@ -1,0 +1,5 @@
+package domain
+
+type (
+	Command struct{}
+)
