+++ bot/domain/discord_messages.go
@@ -1,0 +1,9 @@
+package domain
+
+type (
+	PlainMessage struct{}
+
+	EmbeddedMessage struct{}
+
+	ImageMessage struct{}
+)
