+++ bot/domain/guild.go
@@ -1,0 +1,5 @@
+package domain
+
+type (
+	Guild struct{}
+)
