+++ bot/domain/guild_member.go
@@ -1,0 +1,5 @@
+package domain
+
+type (
+	GuildMember struct{}
+)
