--- deployments/kube/deployment.yaml.tmpl
@@ -1,41 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: template-golang-kubernetes
-  labels:
-    component: template-golang-kubernetes
-spec:
-  replicas: MIN_REPLICAS_VALUE
-  minReadySeconds: 10
-  revisionHistoryLimit: 5
-  strategy:
-    type: RollingUpdate
-    rollingUpdate:
-      maxSurge: 1
-      maxUnavailable: 0
-  selector:
-    matchLabels:
-      component: template-golang-kubernetes
-  template:
-    metadata:
-      labels:
-        component: template-golang-kubernetes
-    spec:
-      containers:
-      - name: template-golang-kubernetes
-        image: tedris/template-golang-kubernetes:VERSION_VALUE
-        imagePullPolicy: Always
-        resources:
-          limits:
-            memory: "128Mi"
-            cpu: "500m"
-        ports:
-        - containerPort: 8080
-        env:
-          - name: LOG_LEVEL
-            value: LOG_LEVEL_VALUE
-          - name: SOME_VARIABLE
-            valueFrom:
-              secretKeyRef:
-                name: some-secret-file
-                key: some_variable
