--- deployments/kube/hpa.yaml.tmpl
@@ -1,18 +1,0 @@
-apiVersion: autoscaling/v2beta2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: template-golang-kubernetes
-spec:
-  scaleTargetRef:
-    apiVersion: apps/v1
-    kind: Deployment
-    name: template-golang-kubernetes
-  minReplicas: MIN_REPLICAS_VALUE
-  maxReplicas: MAX_REPLICAS_VALUE
-  metrics:
-  - type: Resource
-    resource:
-      name: cpu
-      target:
-        type: Utilization
-        averageUtilization: 75
