+++ sample.env
@@ -19,3 +19,5 @@ AWS_ACCESS_KEY_ID=AzureDiamond
 AWS_SECRET_ACCESS_KEY=hunter2
 AWS_REGION=us-east-1
 AWS_ENDPOINT=http://localstack:4566
+
+DISCORD_BOT_TOKEN=abcxyz1234
